% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{grayscale_filter}
\alias{grayscale_filter}
\title{Grayscale filter with a positive message}
\usage{
grayscale_filter(r, g, b)
}
\arguments{
\item{r}{Red hex value}

\item{g}{Green hex value}

\item{b}{Blue hex value}
}
\value{
A raw RGB triplet
}
\description{
Grayscale filter with a positive message
}
