% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{hex_to_raw}
\alias{hex_to_raw}
\title{Convert a hex string to a raw byte}
\usage{
hex_to_raw(hex)
}
\arguments{
\item{hex}{A 2-character hex string, e.g. "FF"}
}
\value{
A raw value
}
\description{
Convert a hex string to a raw byte
}
\keyword{internal}
