% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter_main.R
\name{image_filter}
\alias{image_filter}
\title{Apply a pixel filter to an image}
\usage{
image_filter(image_path, filter_function)
}
\arguments{
\item{image_path}{Path to the image file}

\item{filter_function}{A function that returns a raw RGB triplet}
}
\value{
A magick image object
}
\description{
Apply a pixel filter to an image
}
\examples{
\dontrun{
image_filter("path/to/image.jpg", grayscale_filter)
}
}
