% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{magenta_filter}
\alias{magenta_filter}
\title{Magenta binary filter}
\usage{
magenta_filter(r, g, b, cutoff = 127)
}
\arguments{
\item{r, g, b}{RGB hex values}

\item{cutoff}{Brightness cutoff}
}
\description{
Magenta binary filter
}
