% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{hex_to_raw}
\alias{hex_to_raw}
\title{Convert hex string to raw byte}
\usage{
hex_to_raw(hex)
}
\arguments{
\item{hex}{A hex string, e.g., "FF"}
}
\value{
A raw byte
}
\description{
Convert hex string to raw byte
}
\keyword{internal}
